fix(session): complete startup review guarantees
