+++ crates/horizon-ui/src/app/lifecycle.rs
@@ -13,7 +13,6 @@ use super::{HorizonApp, WS_BG_PAD, WS_TITLE_HEIGHT, attention_feed};
 
 const SPEECH_RELEASE_OWNERSHIP_TIMEOUT: Duration = Duration::from_secs(3);
 const SPEECH_POLL_INTERVAL: Duration = Duration::from_millis(100);
-const STARTUP_BOOTSTRAP_FAILURE_REPAINT_INTERVAL: Duration = Duration::from_secs(1);
 
 #[derive(Clone, Copy, Debug, Eq, PartialEq)]
 struct HoldHotkeyTransition {
@@ -251,20 +250,7 @@ impl HorizonApp {
             self.editor_preview_cache.clear();
         }
 
-        if !self.poll_startup_bootstrap() {
-            self.refresh_active_session_lease();
-            let action = super::session::render_loading_view(ctx, self.startup_bootstrap_failure.as_ref());
-            if let Some(action) = action {
-                self.handle_startup_bootstrap_failure(action);
-                ctx.request_repaint();
-                return false;
-            }
-            let repaint_after = if self.startup_bootstrap_failure.is_some() {
-                STARTUP_BOOTSTRAP_FAILURE_REPAINT_INTERVAL
-            } else {
-                Duration::from_millis(16)
-            };
-            ctx.request_repaint_after(repaint_after);
+        if !self.prepare_startup_bootstrap(ctx) {
             return false;
         }
 
@@ -963,14 +949,11 @@ mod tests {
     use std::time::{Duration, Instant};
 
     use egui::Context;
-    use horizon_core::{
-        Config, HorizonHome, PanelId, RuntimeState, SessionLease, SessionOpenDisposition, SessionStore, StartupDecision,
-    };
+    use horizon_core::{Config, HorizonHome, PanelId, RuntimeState, SessionStore, StartupDecision};
     use tempfile::TempDir;
 
     use super::{
-        HoldHotkeyTransition, HorizonApp, SPEECH_RELEASE_OWNERSHIP_TIMEOUT, STARTUP_BOOTSTRAP_FAILURE_REPAINT_INTERVAL,
-        SpeechActivity, hold_hotkey_transition,
+        HoldHotkeyTransition, HorizonApp, SPEECH_RELEASE_OWNERSHIP_TIMEOUT, SpeechActivity, hold_hotkey_transition,
     };
     use crate::app::HeldSpeechBinding;
     use crate::input;
@@ -995,29 +978,6 @@ mod tests {
         (temp, app)
     }
 
-    fn persistent_test_app() -> (TempDir, HorizonApp, std::path::PathBuf) {
-        let temp = tempfile::tempdir().expect("temp dir");
-        let config_path = temp.path().join("config.yaml");
-        let home = HorizonHome::from_root(temp.path().join(".horizon"));
-        let session_store = SessionStore::new(home.clone(), config_path.clone());
-        let session = session_store
-            .create_session_from_runtime(RuntimeState::default())
-            .expect("create persistent session");
-        let lease_path = home.session_lease_path(&session.session_id);
-        let app = HorizonApp::new_with_egui_context(
-            &Context::default(),
-            &Config::default(),
-            config_path,
-            session_store,
-            StartupDecision::Open {
-                disposition: SessionOpenDisposition::New,
-                session: Box::new(session),
-            },
-            input::ObservedKeyboardInputs::default(),
-        );
-        (temp, app, lease_path)
-    }
-
     #[test]
     fn finalize_frame_requests_repaint_when_theme_application_is_deferred() {
         let ctx = Context::default();
@@ -1032,61 +992,6 @@ mod tests {
         app.finalize_frame(&ctx, false, 0, 0);
 
         assert!(repaint_requests.load(Ordering::Relaxed) > 0);
-    }
-
-    #[test]
-    fn failed_startup_bootstrap_keeps_a_slow_repaint() {
-        let ctx = Context::default();
-        let (_temp, mut app) = test_app();
-        app.startup_bootstrap_failure = Some(super::super::StartupBootstrapFailure::WorkerDisconnected);
-
-        let mut repaint_delay = Duration::ZERO;
-        for _ in 0..8 {
-            let output = ctx.run(egui::RawInput::default(), |ctx| {
-                assert!(!app.prepare_frame(ctx));
-            });
-            repaint_delay = output
-                .viewport_output
-                .get(&egui::ViewportId::ROOT)
-                .expect("root viewport output")
-                .repaint_delay;
-            if !repaint_delay.is_zero() {
-                break;
-            }
-        }
-
-        assert!(!repaint_delay.is_zero());
-        assert!(repaint_delay <= STARTUP_BOOTSTRAP_FAILURE_REPAINT_INTERVAL);
-    }
-
-    #[test]
-    fn failed_startup_bootstrap_refreshes_a_persistent_session_lease() {
-        let ctx = Context::default();
-        let (_temp, mut app, lease_path) = persistent_test_app();
-        let before: SessionLease =
-            serde_yaml::from_str(&std::fs::read_to_string(&lease_path).expect("read acquired lease"))
-                .expect("parse acquired lease");
-        std::thread::sleep(Duration::from_millis(2));
-        app.active_session
-            .as_mut()
-            .expect("active persistent session")
-            .last_lease_refresh = None;
-        app.startup_bootstrap_failure = Some(super::super::StartupBootstrapFailure::WorkerDisconnected);
-
-        let _ = ctx.run(egui::RawInput::default(), |ctx| {
-            assert!(!app.prepare_frame(ctx));
-        });
-
-        let after: SessionLease =
-            serde_yaml::from_str(&std::fs::read_to_string(lease_path).expect("read refreshed lease"))
-                .expect("parse refreshed lease");
-        assert!(after.last_heartbeat_at > before.last_heartbeat_at);
-        assert!(
-            app.active_session
-                .as_ref()
-                .and_then(|session| session.last_lease_refresh)
-                .is_some()
-        );
     }
 
     #[cfg(feature = "speech")]
